+++ ChangeLog.rst
@@ -4,6 +4,8 @@ HEAD
     ``auth_seq_id_map``. The target of this mapping can either be an
     author-provided residue number (as previously) or a 2-element tuple
     containing this number and an insertion code.
+  - :class:`ihm.AsymUnit` now allows the PDB or author-provided strand/chain ID
+    to be different from the regular ID.
   - Bugfix: if two :class:`ihm.Dictionary` objects both contain information
     about a given category, adding the two dictionaries together now combines
     the category information, rather than just using that from one dictionary.